requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.4
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
filelock==3.11.0
gunicorn==20.1.0
h11==0.14.0
huggingface-hub==0.13.4
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
loguru==0.6.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
pinecone-client==2.2.1
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sympy==1.11.1
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
